+++ kffi-android-native/src/main/cpp/kffi_upcall.h
@@ -4,9 +4,13 @@
 #include <jni.h>
 
 /*
- * kffi Android upcall engine: libffi closures that let a C library invoke a
- * Kotlin static dispatcher from an arbitrary thread. See kffi_upcall.c.
+ * kffi Android upcall engine: dynamically typed libffi closures that let a C
+ * library invoke a Kotlin static dispatcher from an arbitrary thread. See
+ * kffi_upcall.c.
  */
+
+#define KFFI_UPCALL_ALLOCATE_JNI_SIGNATURE \
+    "(Ljava/lang/Class;Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;)J"
 
 /* Called once from JNI_OnLoad. */
 void kffi_upcall_init(JavaVM *vm);
@@ -17,7 +21,8 @@ void kffi_upcall_init(JavaVM *vm);
  * JNINativeMethod table for org/graphiks/kffi/engine/UpcallEngine.
  */
 JNIEXPORT jlong JNICALL Java_org_graphiks_kffi_engine_UpcallEngine_allocateTrampoline(
-    JNIEnv *env, jclass cls, jclass dispatcherClass, jstring dispatchMethod, jstring dispatchSig);
+    JNIEnv *env, jclass cls, jclass dispatcherClass, jstring dispatchMethod,
+    jstring dispatchJvmSignature, jstring dispatchAbiSignature);
 JNIEXPORT void JNICALL Java_org_graphiks_kffi_engine_UpcallEngine_freeTrampoline(
     JNIEnv *env, jclass cls, jlong address);